Fix floating-point evaluation of everything.
It's now much faster and can be as accurate as you want. Unfortunately
the exact expressions are no longer calculated which was cool from a
mathematical standpoint but it was too time-consuming and the
expressions got ugly very quickly. In the process of changing to
floating-point, all the simplify() methods were removed.